review: move datadog related utils in a specific file
