+++ apps/ledger-live-mobile/src/screens/Portfolio/index.tsx
@@ -50,7 +50,7 @@ import type { Feature_LlmMmkvMigration } from "@ledgerhq/types-live";
 import { DdRum } from "@datadog/mobile-react-native";
 import { getAccountCurrency } from "@ledgerhq/live-common/account/index";
 import { PORTFOLIO_VIEW_ID, TOP_CHAINS } from "~/utils/constants";
-import { buildFeatureFlagTags } from "~/utils";
+import { buildFeatureFlagTags } from "~/utils/datadogUtils";
 
 type NavigationProps = BaseComposite<
   StackNavigatorProps<WalletTabNavigatorStackParamList, ScreenName.Portfolio>
